requirements.txt
streamlit>=1.32.0
streamlit-folium>=0.18.0

folium>=0.15.0
requests>=2.31.0

torch>=2.0.0
torchvision>=0.15.0

ultralytics>=8.1.0

numpy>=1.23.0
pandas>=1.5.0
Pillow>=9.5.0

opencv-python-headless>=4.8.0

osmnx>=1.8.0
networkx>=3.1
shapely>=2.0.0
geopy>=2.4.0
